refactor: CalendarManagerr를 CalendarViewModel에 통합
