+++ EST_Trip/App/SceneDelegate.swift
@@ -16,7 +16,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
         
-        let root = FeatureFactory.makeSearch()
+        let root = FeatureFactory.makeCalendar()
         let nav = UINavigationController(rootViewController: root)
         window.rootViewController = nav
         window.makeKeyAndVisible()
